requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
Authlib==1.6.1
bandit==1.8.6
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dj-database-url==3.0.1
Django==5.0.4
django-axes==8.0.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dparse==0.6.4
drf-yasg==1.21.10
filelock==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
nltk==3.9.1
packaging==25.0
pbr==6.1.1
pillow==10.4.0
psutil==6.1.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.6.0
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
stevedore==5.4.1
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
psycopg2-binary==2.9.9
